moved Protocol.java to common package
