+++ specs/Naludrag/client/Client.java
@@ -1,5 +1,7 @@
 package client;
 
+import common.Protocol;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
