+++ specs/Naludrag/common/Protocol.java
@@ -1,4 +1,4 @@
-package client;
+package common;
 
 /**
  *
